refactor(render): extract tiled image geometry into OpenSeadragonUtils
Move getTiledImageTransform out of OpenSeadragonRendererBase into
OpenSeadragonUtils as a pure helper taking the object and layer
transforms, mirroring WebGLUtils. The OSD/WebGL parity tests move along
with it and now exercise the shipped functions instead of copies of
them, which is what they were doing in TransformUtils.test.ts.

Take two similarity transforms in the WebGLUtils matrix helpers instead
of a transform and a whole layer.

Compare against getBoundsNoRotate() when updating a tiled image, as
getBounds() rotates around the image center and thus never matches the
position of a rotated tiled image, triggering a redundant write on every
update.
