+++ packages/@tissuumaps-core/src/utils/TransformUtils.test.ts
@@ -1,7 +1,6 @@
-import { mat3, vec2 } from "gl-matrix";
+import { mat3 } from "gl-matrix";
 import { describe, expect, it } from "vitest";
 
-import { identityTransform } from "../model/constants";
 import type { SimilarityTransform } from "../model/primitives";
 import { TransformUtils } from "./TransformUtils";
 
@@ -269,303 +268,5 @@ describe("TransformUtils", () => {
       expect(result.width).toBeCloseTo(6);
       expect(result.height).toBeCloseTo(6);
     });
-  });
-
-  describe("tiled image geometry (similarity matrix composition)", () => {
-    const contentSize = { x: 100, y: 80 };
-
-    /**
-     * Reproduces the OSD tiled-image geometry computation done in
-     * OpenSeadragonController._updateTiledImageGeometry: compose the
-     * data → layer → world similarity matrices (including flip), then
-     * decompose around the image center to match OSD's
-     * flip/rotation-around-image-center convention.
-     */
-    function computeGeometry(
-      transform: SimilarityTransform,
-      layerTransform: SimilarityTransform,
-      contentSize: { x: number; y: number },
-    ): {
-      flip: boolean;
-      width: number;
-      rotation: number;
-      x: number;
-      y: number;
-    } {
-      const m = mat3.create();
-      mat3.multiply(
-        m,
-        TransformUtils.toSimilarityMatrix(layerTransform),
-        TransformUtils.toSimilarityMatrix(transform),
-      );
-      const { flip, scale, rotation, translation } =
-        TransformUtils.fromSimilarityMatrix(m, {
-          center: { x: contentSize.x / 2, y: contentSize.y / 2 },
-        });
-      return {
-        flip,
-        width: contentSize.x * scale,
-        rotation,
-        x: translation.x,
-        y: translation.y,
-      };
-    }
-
-    /**
-     * Simulate where OSD places the four image corners given the returned
-     * geometry parameters (flip, width, rotation, x, y).
-     *
-     * OSD renders a tiled image as:
-     *   1. Place at (x, y) with the given width (height = width * aspect)
-     *   2. If flipped, mirror tile content horizontally around image center
-     *   3. Rotate around image center by `rotation` degrees
-     *
-     * We reverse-engineer corner positions from those parameters.
-     */
-    function osdCorners(
-      geom: {
-        flip: boolean;
-        width: number;
-        rotation: number;
-        x: number;
-        y: number;
-      },
-      contentSize: { x: number; y: number },
-    ): vec2[] {
-      const aspect = contentSize.y / contentSize.x;
-      const w = geom.width;
-      const h = w * aspect;
-      // Corners before rotation, relative to position
-      let corners: vec2[] = [
-        vec2.fromValues(0, 0),
-        vec2.fromValues(w, 0),
-        vec2.fromValues(w, h),
-        vec2.fromValues(0, h),
-      ];
-      if (geom.flip) {
-        // OSD flips around center: x' = w - x
-        corners = corners.map((c) => vec2.fromValues(w - c[0], c[1]));
-      }
-      const cx = w / 2;
-      const cy = h / 2;
-      const rad = (geom.rotation * Math.PI) / 180;
-      const cos = Math.cos(rad);
-      const sin = Math.sin(rad);
-      return corners.map((c) => {
-        const dx = c[0] - cx;
-        const dy = c[1] - cy;
-        return vec2.fromValues(
-          geom.x + cx + dx * cos - dy * sin,
-          geom.y + cy + dx * sin + dy * cos,
-        );
-      });
-    }
-
-    /**
-     * Compute where the WebGL path places the same four corners using
-     * the data → world matrix (same logic as WebGLControllerBase).
-     */
-    function webglCorners(
-      transform: SimilarityTransform,
-      layerTransform: SimilarityTransform,
-      contentSize: { x: number; y: number },
-    ): vec2[] {
-      const dataToLayer = TransformUtils.toSimilarityMatrix(transform);
-      const layerToWorld = TransformUtils.toSimilarityMatrix(layerTransform);
-      const m = mat3.create();
-      mat3.multiply(m, layerToWorld, dataToLayer);
-      const corners = [
-        vec2.fromValues(0, 0),
-        vec2.fromValues(contentSize.x, 0),
-        vec2.fromValues(contentSize.x, contentSize.y),
-        vec2.fromValues(0, contentSize.y),
-      ];
-      return corners.map((c) => {
-        const out = vec2.create();
-        vec2.transformMat3(out, c, m);
-        return out;
-      });
-    }
-
-    it("identity transforms produce correct geometry", () => {
-      const geom = computeGeometry(
-        identityTransform,
-        identityTransform,
-        contentSize,
-      );
-      expect(geom.flip).toBe(false);
-      expect(geom.width).toBeCloseTo(contentSize.x);
-      expect(geom.rotation).toBeCloseTo(0);
-      expect(geom.x).toBeCloseTo(0);
-      expect(geom.y).toBeCloseTo(0);
-    });
-
-    it("applies data scale", () => {
-      const transform: SimilarityTransform = {
-        flip: false,
-        scale: 2,
-        rotation: 0,
-        translation: { x: 0, y: 0 },
-      };
-      const geom = computeGeometry(transform, identityTransform, contentSize);
-      expect(geom.width).toBeCloseTo(200);
-      expect(geom.rotation).toBeCloseTo(0);
-    });
-
-    it("flip-only produces correct position shift", () => {
-      const transform: SimilarityTransform = {
-        flip: true,
-        scale: 1,
-        rotation: 0,
-        translation: { x: 0, y: 0 },
-      };
-      const geom = computeGeometry(transform, identityTransform, contentSize);
-      expect(geom.flip).toBe(true);
-      expect(geom.width).toBeCloseTo(100);
-      // Position must shift left by width to flip around left edge
-      expect(geom.x).toBeCloseTo(-100);
-      expect(geom.y).toBeCloseTo(0);
-    });
-
-    it("layer-level flip XORs with data flip", () => {
-      const transform: SimilarityTransform = {
-        flip: true,
-        scale: 1,
-        rotation: 0,
-        translation: { x: 0, y: 0 },
-      };
-      const layerTransform: SimilarityTransform = {
-        flip: true,
-        scale: 1,
-        rotation: 0,
-        translation: { x: 0, y: 0 },
-      };
-      const geom = computeGeometry(transform, layerTransform, contentSize);
-      // flip XOR flip = no flip
-      expect(geom.flip).toBe(false);
-    });
-
-    it.each([
-      {
-        name: "scale + rotation, no flip",
-        transform: {
-          flip: false,
-          scale: 2,
-          rotation: 45,
-          translation: { x: 10, y: 5 },
-        },
-        layerTransform: identityTransform,
-      },
-      {
-        name: "flip + rotation",
-        transform: {
-          flip: true,
-          scale: 1,
-          rotation: 30,
-          translation: { x: 0, y: 0 },
-        },
-        layerTransform: identityTransform,
-      },
-      {
-        name: "flip + scale + rotation + translation",
-        transform: {
-          flip: true,
-          scale: 2,
-          rotation: 60,
-          translation: { x: -5, y: 10 },
-        },
-        layerTransform: identityTransform,
-      },
-      {
-        name: "layer flip + data no-flip",
-        transform: {
-          flip: false,
-          scale: 1.5,
-          rotation: 0,
-          translation: { x: 3, y: 7 },
-        },
-        layerTransform: {
-          flip: true,
-          scale: 1,
-          rotation: 0,
-          translation: { x: 0, y: 0 },
-        },
-      },
-      {
-        name: "both flip + layer rotation",
-        transform: {
-          flip: true,
-          scale: 1,
-          rotation: 0,
-          translation: { x: 0, y: 0 },
-        },
-        layerTransform: {
-          flip: true,
-          scale: 1,
-          rotation: 90,
-          translation: { x: 0, y: 0 },
-        },
-      },
-      {
-        name: "layer scale+rotation + data translation",
-        transform: {
-          flip: false,
-          scale: 1.5,
-          rotation: 0,
-          translation: { x: 20, y: 10 },
-        },
-        layerTransform: {
-          flip: false,
-          scale: 2,
-          rotation: 30,
-          translation: { x: 5, y: 3 },
-        },
-      },
-      {
-        name: "layer scale+rotation + data scale+rotation+translation",
-        transform: {
-          flip: false,
-          scale: 2,
-          rotation: 15,
-          translation: { x: -5, y: 10 },
-        },
-        layerTransform: {
-          flip: false,
-          scale: 3,
-          rotation: 45,
-          translation: { x: 10, y: -7 },
-        },
-      },
-      {
-        name: "flip + layer scale+rotation+translation + data translation",
-        transform: {
-          flip: true,
-          scale: 1.5,
-          rotation: 20,
-          translation: { x: 8, y: -3 },
-        },
-        layerTransform: {
-          flip: false,
-          scale: 2,
-          rotation: 60,
-          translation: { x: -4, y: 12 },
-        },
-      },
-    ] as {
-      name: string;
-      transform: SimilarityTransform;
-      layerTransform: SimilarityTransform;
-    }[])(
-      "OSD corners match WebGL corners: $name",
-      ({ transform, layerTransform }) => {
-        const geom = computeGeometry(transform, layerTransform, contentSize);
-        const osd = osdCorners(geom, contentSize);
-        const webgl = webglCorners(transform, layerTransform, contentSize);
-        for (let i = 0; i < 4; i++) {
-          expect(osd[i]![0]).toBeCloseTo(webgl[i]![0], 3);
-          expect(osd[i]![1]).toBeCloseTo(webgl[i]![1], 3);
-        }
-      },
-    );
   });
 });
